+++ database/seeders/DatabaseSeeder.php
@@ -17,8 +17,11 @@ class DatabaseSeeder extends Seeder
     {
         $this->call([
             PeranSeeder::class,
+            GuruSeeder::class,
+            SiswaSeeder::class,
             MateriSeeder::class,
             KuisSeeder::class,
+            PercobaanKuisSeeder::class,
         ]);
     }
 }